additional code organization, renaming, and test stub creation
