+++ flutter_app/lib/domains/messages/SocketMessage.dart
