+++ flutter_app/lib/domains/messages/incoming-messages/SentMessageMessage.dart
