+++ flutter_app/lib/domains/messages/outgoing-messages/chat-message/SendMessageMessage.dart
@@ -1,6 +1,6 @@
 import 'package:flutterapp/domains/messages/outgoing-messages/chat-message/SendMessageContent.dart';
 
-import '../../Message.dart';
+import '../../SocketMessage.dart';
 
 class SendMessageMessage extends SocketMessage {
   String type = 'sendMessage';
