+++ flutter_app/lib/domains/messenger/Messenger.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutterapp/domains/messages/Message.dart';
+import 'package:flutterapp/domains/messages/SocketMessage.dart';
 import 'package:web_socket_channel/io.dart';
 
 class Messenger {
