+++ flutter_app/test/domains/messages/SocketMessage_test.dart
@@ -1,0 +1,11 @@
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  group('socket message unit tests', () {
+    group('constructor fromString', () {
+      test('that each supported message can be created', () {
+        // TODO: write these tests
+      });
+    });
+  });
+}
