+++ flutter_app/test/domains/messages/incoming-messages/ReceivedMessageUtility_test.dart
@@ -1,0 +1,11 @@
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  group('received message utility unit tests', () {
+    group('fromString', () {
+      test('that each message type can be interpreted', () {
+        // TODO: write tests for this
+      });
+    });
+  });
+}
